+++ test/css.js
@@ -1,21 +1,19 @@
 import validator from 'stylelint-csstree-validator';
-import lessSyntax from 'postcss-less';
-import scssSyntax from 'postcss-scss';
 import ruleTester from './utils/tester.js';
 
 const { rule, ruleName, messages } = validator;
 const css = ruleTester(rule, ruleName);
-const less = ruleTester(rule, ruleName, {
-    postcssOptions: { syntax: lessSyntax }
-});
-const sass = ruleTester(rule, ruleName, {
-    postcssOptions: { syntax: scssSyntax }
-});
-const invalidValue = (prop, line, column) => {
-    return invalid(messages.invalidValue(prop), line, column);
+const unknownAtrule = (atrule, line, column) => {
+    return invalid(messages.unknownAtrule(atrule), line, column);
 };
 const invalidPrelude = (atrule, line, column) => {
     return invalid(messages.invalidPrelude(atrule), line, column);
+};
+const unknownProperty = (atrule, line, column) => {
+    return invalid(messages.unknownProperty(atrule), line, column);
+};
+const invalidValue = (prop, line, column) => {
+    return invalid(messages.invalidValue(prop), line, column);
 };
 const invalid = (message, line, column) => {
     if (typeof line !== 'number' && typeof column !== 'number') {
@@ -36,7 +34,10 @@ css(null, function(tr) {
     tr.notOk('.foo { color: 1 }', invalidValue('color'));
     tr.notOk('.foo { color: #12345 }', invalidValue('color'));
     tr.notOk('.foo { color: &a }', messages.parseError('&a'));
-    tr.notOk('.foo { baz: 123 }', 'Unknown property `baz`');
+    tr.notOk('.foo { baz: 123 }', unknownProperty('baz'));
+    tr.notOk('@plugin "my-plugin"', unknownAtrule('plugin'));
+    tr.notOk('@test;', unknownAtrule('test'));
+    tr.notOk('.foo { color: @foo }', messages.parseError('@foo'));
 });
 
 // loc
@@ -47,7 +48,7 @@ css(null, function(tr) {
 
 // atrules
 css(null, function(tr) {
-    tr.notOk('  @unknown {}', invalid('Unknown at-rule `@unknown`', 1, 3));
+    tr.notOk('  @unknown {}', unknownAtrule('unknown', 1, 3));
     tr.notOk('  @media ??? {}', invalidPrelude('media', 1, 10));
 });
 css({ ignoreAtrules: ['unknown', 'import'] }, function(tr) {
@@ -60,73 +61,12 @@ css({ atrules: false }, function(tr) {
     tr.ok('@media ??? {}');
 });
 
-// ignore values with less extenstions
-less(null, function(tr) {
-    // variables
-    tr.ok('.foo { color: @var }');
-    tr.ok('.foo { color: @@var }');
-    tr.ok('.foo { color: @ }', messages.parseError('@'));
-    tr.ok('.foo { color: @123 }', messages.parseError('@123'));
-    tr.ok('.foo { color: @@@var }', messages.parseError('@@@var'));
-
-    // escaping
-    tr.ok('.foo { color: ~"test" }');
-    tr.ok('.foo { color: ~\'test\' }');
-    tr.notOk('.foo { color: ~ }', messages.parseError('~'));
-    tr.notOk('.foo { color: ~123 }', messages.parseError('~123'));
-
-    // interpolation
-    tr.ok('.foo { @{property}: 1 }');
-    tr.ok('.foo { test-@{property}: 1 }');
-    tr.ok('.foo { @{property}-test: 1 }');
-
-    // standalone var declarations
-    tr.ok('@foo: 2');
-
-    // namespaces
-    tr.ok('.foo { color: #outer > .inner(); }');   // deprecated
-    tr.ok('.foo { color: #outer> .inner(); }');    // deprecated
-    tr.ok('.foo { color: #outer >.inner(); }');    // deprecated
-    tr.ok('.foo { color: #outer>.inner(); }');     // deprecated
-    tr.ok('.foo { color: #outer .inner(); }');     // deprecated
-    tr.ok('.foo { color: #outer.inner(); }');      // preferred
-    tr.ok('.foo { color: #outer.inner(1 + 2); }'); // preferred
-});
-
-// ignore values with sass extenstions
-sass(null, function(tr) {
-    // variables
-    tr.ok('.foo { color: $red }');
-    tr.ok('.foo { color: $ }', messages.parseError('$'));
-    tr.ok('.foo { color: $123 }', messages.parseError('$123'));
-    tr.ok('.foo { color: $$123 }', messages.parseError('$$123'));
-
-    // modulo operator
-    tr.ok('.foo { color: 3 % 6 }');
-
-    // interpolation
-    tr.ok('.foo { color: #{$var} }');
-    tr.ok('.foo { color: #{1 + 2} }');
-    tr.ok('.foo { max-height: calc(100vh - #{$navbar-height}); }');
-    tr.ok('.foo { #{$property}: 1 }');
-    tr.ok('.foo { test-#{$property}: 1 }');
-    tr.ok('.foo { #{$property}-test: 1 }');
-
-    // standalone var declarations
-    tr.ok('$foo: 1');
-
-    // namespace
-    tr.ok('.foo { color: fonts.$font-family-text; }');
-    tr.ok('.foo { color: fonts.foo(); }');
-    tr.ok('.foo { color: fonts.foo(1 + 2); }');
-});
-
 // should ignore properties from `ignore` list
 css({ ignoreProperties: ['foo', 'bar'] }, function(tr) {
     tr.ok('.foo { foo: 1 }');
     tr.ok('.foo { bar: 1 }');
     tr.ok('.foo { BAR: 1 }');
-    tr.notOk('.foo { baz: 1 }', 'Unknown property `baz`');
+    tr.notOk('.foo { baz: 1 }', unknownProperty('baz'));
 });
 
 // should ignore by ignoreValue pattern
@@ -137,8 +77,8 @@ css({ ignoreValue: '^patternToIgnore$|=', ignoreProperties: ['bar'] }, function(
     tr.ok('.foo { bar: notMatchingPattern }');
     tr.ok('.foo { color: alpha(opacity=1) }');
     tr.notOk('.foo { color: notMatchingPattern }', invalidValue('color'));
-    tr.notOk('.foo { foo: patternToIgnore }', invalid('Unknown property `foo`', 1, 8));
-    tr.notOk('.foo { foo: notMatchingPattern }', 'Unknown property `foo`');
+    tr.notOk('.foo { foo: patternToIgnore }', unknownProperty('foo', 1, 8));
+    tr.notOk('.foo { foo: notMatchingPattern }', unknownProperty('foo'));
 });
 
 // extend dictionary
@@ -159,7 +99,7 @@ css({
     tr.ok('.foo { foo: my-fn(10%) }');
     tr.ok('.foo { foo: my-fn(0) }');
     tr.ok('.bar { bar: my-fn(10px) }');
-    tr.notOk('.baz { baz: my-fn(10px) }', 'Unknown property `baz`');
+    tr.notOk('.baz { baz: my-fn(10px) }', unknownProperty('baz'));
     tr.ok('.qux { qux: 10px }');
     tr.notOk('.foo { color: my-fn(10px) }', invalidValue('color'));
     tr.ok('.foo { color: darken(white, 5%) }');
@@ -195,14 +135,26 @@ css({
     tr.ok('@font-face { font-display: ignore-this }');
     tr.ok('@font-face { ignore-descriptor: foo }');
     tr.notOk('@font-face { font-display: ignore-that }', invalidValue('font-display', 1, 28));
-    tr.notOk('.foo { font-display: swap }', invalid('Unknown property `font-display`', 1, 8));
-    tr.notOk('  @not-import url("foo.css");', invalid('Unknown at-rule `@not-import`', 1, 3));
-    // tr.notOk('  @-unknown-import url("foo.css");', invalid('Unknown at-rule `@-unknown-import`', 1, 3));
-    tr.notOk('  @import { color: red }', invalid('At-rule `@import` should contain a prelude', 1, 11));
+    tr.notOk('.foo { font-display: swap }', invalid(`Unknown property \`font-display\` (${ruleName})`, 1, 8));
+    tr.notOk('  @not-import url("foo.css");', unknownAtrule('not-import', 1, 3));
+    // tr.notOk('  @-unknown-import url("foo.css");', unknownAtrule("-")known-mport`', 1, 3));
+    tr.notOk('  @import { color: red }', invalid(`At-rule \`@import\` should contain a prelude (${ruleName})`, 1, 11));
     tr.notOk('  @import url("foo.css") .a', invalidPrelude('import', 1, 26));
-    tr.notOk('  @font-face xx {}', invalid('At-rule `@font-face` should not contain a prelude', 1, 14));
+    tr.notOk('  @font-face xx {}', invalid(`At-rule \`@font-face\` should not contain a prelude (${ruleName})`, 1, 14));
     tr.notOk('  @font-face { font-display: foo }', invalidValue('font-display', 1, 30));
-    tr.notOk('  @font-face { font-displa: block }', invalid('Unknown at-rule descriptor `font-displa`', 1, 16));
-    tr.notOk('  @foo { color: ref }', [invalid('Unknown at-rule `@foo`', 1, 3)]);
+    tr.notOk('  @font-face { font-displa: block }', invalid(`Unknown at-rule descriptor \`font-displa\` (${ruleName})`, 1, 16));
+    tr.notOk('  @foo { color: ref }', [unknownAtrule('foo', 1, 3)]);
     tr.notOk('  @media zzz zzz { color: ref }', [invalidPrelude('media', 1, 14)]);
+});
+
+// options varnings
+css({
+    ignore: ['ignore-descriptor']
+}, function(tr) {
+    tr.notOk('a {}', `Invalid value "ignore-descriptor" for option "ignore" of rule "${ruleName}"`);
+});
+css({
+    syntaxExtensions: ['sass', 'xxx']
+}, function(tr) {
+    tr.notOk('a {}', `Invalid value "xxx" for option "syntaxExtensions" of rule "${ruleName}"`);
 });